Update OpenGL funcs

+++ Exemples/OPENGL_IN_SDL/OPENGL_IN_SDL/Code/Sprite.cs
@@ -286,13 +286,13 @@ namespace OPENGL_IN_SDL
             int _vertexBufferObject;
             _vertexBufferObject = GL.GenBuffer();
             GL.BindBuffer(BufferTarget.ArrayBuffer, _vertexBufferObject);
-            GL.BufferData(BufferTarget.ArrayBuffer, Vertices.Length * sizeof(float), Vertices, BufferUsageHint.StaticDraw);
+            GL.BufferData(BufferTarget.ArrayBuffer, (IntPtr)(Vertices.Length * sizeof(float)), Vertices, BufferUsageHint.StaticDraw);
             vertexBufferObject.Add(_vertexBufferObject);
 
             int _elementBufferObject;
             _elementBufferObject = GL.GenBuffer();
             GL.BindBuffer(BufferTarget.ElementArrayBuffer, _elementBufferObject);
-            GL.BufferData(BufferTarget.ElementArrayBuffer, Indices.Length * sizeof(uint), Indices, BufferUsageHint.StaticDraw);
+            GL.BufferData(BufferTarget.ElementArrayBuffer, (IntPtr)(Indices.Length * sizeof(uint)), Indices, BufferUsageHint.StaticDraw);
             elementBufferObject.Add(_elementBufferObject);
 
             Shader.asset_shader.Use();
